Create Understanding control flow
Understanding control flow
Let's put that understanding of control flow to the test. You have a team member who needs some help writing a simple procedural program that will print out all of the names on her list, names_list, that start with the letter "A". Can you help her write a code snippet that achieves this?

Instructions

Fill out the code given to loop through the list of names and print only those names starting with the letter "A".