Create Changing programming paradigms 2
Changing programming paradigms
As sometimes happens, you've started out on a project using one paradigm and now realized that it wasn't quite the right one for the task. You've currently got a simple procedural program for storing automobile sales information that is going to need to be converted to an object-oriented program so that it can be extended more easily.

Instructions 2/2

Finish the implementation of the car purchase logic by filling in the missing details so that the logic matches the original program.